restructure run code, preparing for EditRunner
