+++ src/cli/editCommand.ts
@@ -1,7 +1,7 @@
 import program, { Command } from 'commander';
 import { loadConfig } from '../config';
 import { getCodePath } from './getCodePath';
-import { runEdit } from '../run/runEdit';
+import { EditRunner } from '../run/EditRunner';
 
 export const buildEditCommand = (): program.Command => {
   const command = new Command('edit')
@@ -9,27 +9,21 @@ export const buildEditCommand = (): program.Command => {
     .storeOptionsAsProperties(false)
     .description('📝 edit a test')
     .action(async () => {
-      try {
-        const [name] = command.args;
+      const [name] = command.args;
 
-        const config = loadConfig();
+      const config = loadConfig();
 
-        const codePath = await getCodePath({
-          rootDir: config.rootDir,
-          name,
-          useTypeScript: config.useTypeScript,
-        });
+      const codePath = await getCodePath({
+        rootDir: config.rootDir,
+        name,
+        useTypeScript: config.useTypeScript,
+      });
 
-        runEdit({
-          codePath,
-          config,
-          isScript: !codePath.includes('test'),
-        });
-
-        process.exit(0);
-      } catch (e) {
-        process.exit(1);
-      }
+      await EditRunner.start({
+        codePath,
+        config,
+        isScript: !codePath.includes('test'),
+      });
     });
 
   return command;
