+++ test/run/runTests.test.ts
@@ -1,38 +1,38 @@
-import { buildArguments, runJest } from '../../src/run/runJest';
 import { join, resolve } from 'path';
+import { buildJestArguments, runTests } from '../../src/run/runTests';
 
 const rootDir = join(__dirname, '../.qawolf');
 
-describe('buildArguments', () => {
+describe('buildJestArguments', () => {
   describe('provided args', () => {
     it('passes them through', () => {
       const args = ['--sup=yo', '--hello=there'];
-      const builtArgs = buildArguments({ args });
+      const builtArgs = buildJestArguments({ args });
       expect(builtArgs.slice(builtArgs.length - 2)).toEqual(args);
     });
   });
 
   describe('config', () => {
     it('uses config when specified', () => {
-      expect(buildArguments({ config: '{}' })).toContain('--config="{}"');
+      expect(buildJestArguments({ config: '{}' })).toContain('--config="{}"');
     });
 
     it('does not use config if not specified', () => {
-      const builtArgs = buildArguments({});
+      const builtArgs = buildJestArguments({});
       expect(builtArgs.filter((arg) => arg.includes('--config'))).toEqual([]);
     });
   });
 
   describe('reporters', () => {
     it('uses basic reporter for repl', () => {
-      const args = buildArguments({ repl: true });
+      const args = buildJestArguments({ repl: true });
       expect(args).toContain('--reporters="@qawolf/jest-reporter"');
     });
   });
 
   describe('testTimeout', () => {
     it('not set when one is provided', () => {
-      const args = buildArguments({ args: ['--testTimeout=1000'] });
+      const args = buildJestArguments({ args: ['--testTimeout=1000'] });
       const timeoutArgs = args.filter((a) =>
         a.toLowerCase().includes('timeout'),
       );
@@ -40,21 +40,25 @@ describe('buildArguments', () => {
     });
 
     it('use provided testTimeout', () => {
-      expect(buildArguments({ testTimeout: 10 })).toContain('--testTimeout=10');
+      expect(buildJestArguments({ testTimeout: 10 })).toContain(
+        '--testTimeout=10',
+      );
     });
 
     it('set to 60s by default', () => {
-      expect(buildArguments({})).toContain('--testTimeout=60000');
+      expect(buildJestArguments({})).toContain('--testTimeout=60000');
     });
 
     it('set to 1 hour for repl', () => {
-      expect(buildArguments({ repl: true })).toContain('--testTimeout=3600000');
+      expect(buildJestArguments({ repl: true })).toContain(
+        '--testTimeout=3600000',
+      );
     });
   });
 
   describe('test path', () => {
     it('replaces forward slashes for powershell to work', () => {
-      const args = buildArguments({
+      const args = buildJestArguments({
         testPath: 'C:\\Users\\jon\\Desktop\\qawolf\\.qawolf\\test.test.js',
       });
       expect(args).toContain(
@@ -64,16 +68,16 @@ describe('buildArguments', () => {
   });
 });
 
-describe('runJest', () => {
+describe('runTests', () => {
   it('runs successful test for a relative path', () => {
     expect(() =>
-      runJest({ args: ['scroll'], browsers: ['chromium'], rootDir }),
+      runTests({ args: ['scroll'], browsers: ['chromium'], rootDir }),
     ).not.toThrow();
   });
 
   it('runs successful test for a test path', () => {
     expect(() =>
-      runJest({
+      runTests({
         browsers: ['chromium'],
         rootDir,
         testPath: resolve(join(rootDir, 'scroll.test.js')),
@@ -83,7 +87,7 @@ describe('runJest', () => {
 
   it('throws error for failed test', () => {
     expect(() =>
-      runJest({ args: ['notATest'], browsers: ['chromium'] }),
+      runTests({ args: ['notATest'], browsers: ['chromium'] }),
     ).toThrow();
   });
 });
